fix: resolve test compilation errors for v1.1.0 release
Fixed 3 pre-existing test compilation errors:

1. MetricsCacheTest.kt - Added missing coroutines imports (async, awaitAll)
2. FpsMetricsCollectorTest.kt - Fixed illegal characters in method names
   - Replaced >= with 'at least'
   - Replaced < with 'below'
3. MetricsRepositoryAggregationTest.kt - Added missing coroutines imports

Additional fixes:
- Disabled explicit API mode for test source sets (both modules)
- Fixed assertEquals type mismatch in FpsCalculationTest.kt

All code now compiles successfully.

+++ sysmetrics-overlay/src/test/kotlin/com/sysmetrics/overlay/fps/FpsCalculationTest.kt
@@ -91,7 +91,7 @@ class FpsCalculationTest {
         
         val fps = frameDurationToFps(durationMs)
         
-        assertEquals(60, fps, 1) // Allow 1 fps margin
+        assertEquals(60.0, fps.toDouble(), 1.0) // Allow 1 fps margin
     }
 
     @Test
